Display "Song Folder" or "Group" On SelectMusic and Evaluation
"Yo dude, what pack is this song from?"

Basically the question everyone asks whenever scores are posted. You can
choose between displaying song folder or group on ScreenSelectMusic (but not on ScreenEvaluation). This preference can be found in Simply Love Options; the default is to show the song group. This is very useful if you're sorting the songwheel by anything other than group.

-Folder displays the full folder name of the song selected. Example:
"Eurobeat Is Fantastic/Deja Vu - [Arvin]" Both the name of the pack
folder, and the folder the song is found in are listed. This is great
for looking for specific songfiles, and especially since song folder
names usually have "useful" information in them.

-Group displays only the pack that the selected song is in. This is good
for arcade owners that are concerned about potential "questionable
language" in a songfolder name. I'm pretty sure arcade owners know
already not to put questionable content in their pack folder names.

-ScreenEvaluation will only display group name because I don't think
it's important to have redundant information (since the song folder will
almost always have the name of the song in it, which is displayed
directly below the group information). If there was an interest, it's
super easy to make this configurable in Simply Love options.

-ScreenEvaluation has been refreshed to show an information card below the song banner mimmicking ScreenSelectMusic that shows the song group, BPM (and rate mod), and song length.

-Casual Mode has no UI changes through the use of conditionals

-CourseMode will only show the BPMs in the card below the course banner because length and song group are not really important in this mode

-This also tweaks the long/marathon version bubble ActorMultiVertex to be less obtrusive against the Length information that was nudged down to make space for the song folder text.