Player Profile displayed on SSMW and Evaluation
Player Profile displayed on ScreenSelectMusicWide and ScreenEvaluationStage

There is now a card at the top of the player's side of the screen (on ScreenSelectMusic and ScreenEvaluation common) showing a small number of stats that relate to the player. There is an avatar, counts for the number of stars (quads, tri, double, single) achieved across all songs played by the player's profile, and a readout of the number of times the current highlighted song has been played.

Player profile avatar photos are 80x80px.

Player Avatar/name does not appear in the bottom bar on ScreenSelectMusicWide or ScreenEvaluation because these pieces of information are now in the profile card.

Players that don't have an avatar in their profile will be given a default avatar based on the current VisualStyle.