fixed(?) race condition with song search
the newest build of Outfox caused a race condition to occur when searching for songs in a large music library. By increasing the sleep command, this gives enough time for the inputhandler to catch up to the theme. Previously, activating search would lead to a loss of all input because the aforementioned inputhandler race condition. I don't exactly know what's going on here, but 250ms isn't enough of a difference to me so let's just put a bandaid on things.