Clarify ImmediateContinue and GrooveStats scores
The ImmediateContinue failtype was renamed to "delayed", but other references to it were not renamed. This commit clarifies which failtype must be used in order for GrooveStats scores (and machine scores, hopefully) to be valid.

To be clear here: you must set your failtype to either "Immediate" or "Delayed" in order for your scores to be counted for GrooveStats.