Move stream breakdown to under GraphDisplay
I argue that a player's optionlist contains far less pertinent information than a song's breakdown stats. If a player selects a disqualifiable mod, they'll know because a "disqualified" banner will show. Players who know about those mods should know to disable them without looking for them in their optionslist. As a bit of a compromise, the player's optionlist appears briefly in place of the breakdown text but fades out as the breakdown text fades in.