Rearrange Player Options for better usability
Many Player Options were disorganized and difficult to wrangle for new users. This rearrangement should help new players get a grasp on the options available to them. The options are now also better arranged in a more logical grouping.

Some notable changes are:

-StepStatsExtra are now shown in one row

-converge Gameplay Extras into one option row

-rename "Break UI Elements" to "Break GUI" because there just barely wasn't enough room to prevent overflow

-remove ShowHeldMiss preference in GameplayExtras because it's redundant?

-ActionOnMissedTarget is still useful outside of EventMode with the option to simply dim the score

-Move NoteFieldOffsetX&Y Mods to Uncommon Options

Note: The NoteFieldOffset modifiers are extremely niche options used in very particular usecases where the physical gamepad does not physically align well with the game monitor. These options are also just not useful on a song-to-song basis so it's best to move these options to the Uncommon Options screen.