More Accurate Tactics Judgement Font
The upstream Tactics Judgement Font was a less faithful recreation of this font (the original artwork is too low of a resolution to simply just drop into this theme). I am in favor of replacing the upstream font rather than having both available; anyone who remembers using the Tactics theme should ultimately prefer the better recreation. The following is the comment message provided by @PkGamTV

-------------------

I recently made a pixel perfect fix and port of Mad Matt's In The Groove Tactics judgefont as maru in Stamina Nation was looking for one, found it, but noted there were no white Fantastic windows. So I added them, then tweaked some bits of it to fix some sight flaws the original had like missing pixel fills on the right excellent's dash and being right of center which also some glow cutoff on the right. Finally I enlarged it evenly on both dimensions with no interpolation to keep it's pixels as sharp as possible. People seemed quite happy to see it, so I thought that maybe I should ask if you guys would like to implement it. This one is at 3 times their regular size because that's the closest match to the other existing ones in ITGmania/Simply Love.

The only thing is I don't know what you'd want to name it to fit in with the other Tactics judgefont so I left it as "OldTactics" for now. Like alphabetically maybe it should be "TacticsOG" or "Tactics-Old" or something so it'll pair with it and be more intuitive to find. That's your call too. (Alternately the tactics already implemented could be renamed to something else of course, but that is probably a bad idea.)

Co-Authored-By: PkGamTV <170469263+pkgamtv@users.noreply.github.com>