Visual GUI support for OutFox Online Profile Saving Prompts
Initial implementation of OutFox Online elements. There will be a notification prompt that shows the user that the client has connected to the OutFox Online server, along with notifications for a submitted score, and profile save/sync.

Keep in mind that ScreenSystemOverlay screens and their derivatives do not support OnCommand functionality. In order to get animations to play, you have to use MessageCommands instead.

Co-Authored-By: Jose Varela <jose@josevarela.net>