stepartist textbox more resembles SL 3.95, move PeakNPS/eBPM
-Replaces the stepartist quad with an ActorMultiVertex similar to the "COUNTS AS X ROUNDS" bubble. This asset is also longer, spanning the entire width of DensityGraph (length adjusted smaller to fit in 4:3).

-Used the same animation of the stepartist bubble flying in/out as in SL3.95 .

-Reorder the loading of StepsDisplayList and and PerPlayer so that the the stepartistbubble loads on top of the StepsDisplayList.

-PeakNPS/eBPM no longer appears to be "floating" above the NPS graph. Previously, in 4:3, eBPM was even cut off because it wasn't adapted for the narrower aspect ratio.

I originally wanted to stretch the stepartist bubble to be long enough to reach PeakNPS, but I couldn't figure out the layering conflicts whereby PeakNPS and the NPSgraph are in the same ActorFrame, yet PeakNPS would have to draw higher than the Stepartist bubble while the bubble also draws on top of the NPSgraph. Since draworder doesn't work to fix the layering conflict, the only way to fix this would be to completely merge the two.

eBPM Discussion:

eBPM is equal to the peak NPS in a chart, whereby said NPS is represented by a BPM where 16th notes equaled the same NPS. Peak NPS is a relatively new metric that tells the player what quantization of notes they can expect to encounter at the peak of a chart's difficulty, but is a "fun number" that non-players can easily understand; Peak NPS is far more abstract to skilled Stepmania players. It is much less abstract to the skilled Stepmania player to consider the eBPM and ascertain as to what quantization a chart follows; a 110bpm chart with a max eBPM of 165 tells the player that there are 24th note sections. eBPM also helps in the case of "skittles" charts to tell the player what BPM they will be playing at; many skittles patterns are either very difficult to visually discern because of the lack of quantization colours available in Stepmania (20th notes are a particular painpoint, at this time).

Math:
-If eBPM = BPM/2, then the chart's hardest section contains 8th notes
-If eBPM = BPM, then the chart's hardest section contains 16th notes
-If eBPM = BPM * 1.5, then the chart's hardest section contains 24th notes
-If eBPM = BPM *2, then the chart's hardest section contains 32nd notes.

Disclaimer: Peak eBPM will be incorrect if sections of peak NPS are too short. This was not an issue for NPS readouts. In the future it may be worth looking into parsing a chart based on the quantization, rather than just purely how many notes appear per second.