Upstream bugfix: Stage counter on SSM
There is a longstanding bug that I only just recently found that makes the stage counter code from upstream not work.

Instead of fixing the upstream code, I merged together code from Zmod and the original code that works better than it did before. See the various comments in this commit for context.