(Bugfix) Use functions with SortMenu conditionals
Upstream has a bugfix where conditionals are applied via a function: certain options were not showing up unless SSM/SSMw was refreshed. Using a function in the conditional portion of the SortMenu options ensures that each option is available when first loading the SortMenu.

In the future, if you are having issues with your custom option not being available on first load, use a function call instead.

I separated this commit out from the upstream commit because I want to draw attention to this bugfix. Future re

Co-Authored-By: Crash Cringle <30600688+CrashCringle12@users.noreply.github.com>